Rewriting Scripts to work with New DB structure
Minor edits to both DB and the IngredientStack scripts as the API for
the database gets more fleshed out. Worth noting is that DatabaseEntry
now is a class and has a clone method as well as a constructor that is
special for that clone method.

Mechanics have also been minorly tweaked to make them more Unity Esque.
They are now mono-behaviour objects and need to be created as a
component and then initialised.

Once I get out of this muck with the DB I will likely write a wrapper
(OR SOMEONE ELSE COULD) that will simplify this creation process and
ensure it is always done in the correct area of code!